refactor(Prompts): move display components to display/ subdirectory
