+++ client/src/components/Prompts/display/EmptyPromptPreview.tsx
