+++ client/src/components/Prompts/display/index.ts
@@ -1,0 +1,4 @@
+export { default as PromptDetails } from './PromptDetails';
+export { default as PromptVariables } from './PromptVariables';
+export { default as PromptVersions } from './PromptVersions';
+export { default as EmptyPromptPreview } from './EmptyPromptPreview';
